black bg for meters, font fixups for startup message, new notebook widget (incomplete)
git-svn-id: svn://localhost/trunk/ardour2@178 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/editor_tempodisplay.cc
@@ -217,7 +217,7 @@ Editor::draw_measures ()
 				line = get_time_line ();
 				line->property_x1() = xpos;
 				line->property_x2() = xpos;
-				line->property_y2() = (gdouble) canvas_height;
+				line->property_y2() = canvas_height;
 				line->property_color_rgba() = color;
 				line->raise_to_top();
 				line->show();
